User Task Crud working. Check Repetitive Tasks working with understanded logic.

+++ LegalTrace.DAL/Controllers/ClientControllers/ClientDelete.cs
@@ -14,10 +14,15 @@ namespace LegalTrace.DAL.Controllers.ClientControllers
         public async Task<bool> DeleteClient(int id)
         {
             var client = await _context.Clients.Where(clientAux => clientAux.Id.Equals(id)).FirstOrDefaultAsync();
-            _context.Clients.Remove(client);
-            if (await _context.SaveChangesAsync() > 0) { 
-                return true;
+            if(client != null)
+            {
+                _context.Clients.Remove(client);
+                if (await _context.SaveChangesAsync() > 0)
+                {
+                    return true;
+                }
             }
+            
             return false;
         }
     }
